:recycle: refactor: Adding custom exceptions in service

+++ api-encryption/src/main/java/com/ggomes/api_encryption/services/EncryptService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.ggomes.api_encryption.entities.EncryptEntity;
+import com.ggomes.api_encryption.exceptions.InvalidRequestException;
+import com.ggomes.api_encryption.exceptions.ResourceNotFoundException;
 import com.ggomes.api_encryption.repositories.EncryptRepository;
 
 @Service
@@ -21,14 +23,20 @@ public class EncryptService {
 
     public EncryptEntity findById(Long id) {
         return repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Transaction not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Transaction not found with ID: " + id));
     }
 
     public EncryptEntity save(EncryptEntity transaction) {
+    	if (transaction.getValue() <= 0) {
+            throw new InvalidRequestException("Transaction value must be greater than zero");
+        }
         return repository.save(transaction);
     }
 
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Cannot delete, transaction not found with ID: " + id);
+        }
         repository.deleteById(id);
     }
 }
